Outline per-type parsers to reduce f18 build-time CPU & memory requirements
