+++ lib/parser/user-state.cc
@@ -13,9 +13,6 @@
 // limitations under the License.
 
 #include "user-state.h"
-#include "basic-parsers.h"
-#include "grammar.h"
-#include "openmp-grammar.h"
 #include "parse-state.h"
 #include "stmt-parser.h"
 #include "type-parsers.h"
